[Feat] 입력 값이 부족할 경우 메시지
첫 번째 입력값, 두 번째 입력값, 연산자 중 하나라도 입력되지 않고 등호 입력시 오류 메시지 프린트
나머지 값 구할 때, Double 내 truncatingRemainder 메소드 이용

+++ Calculator.playground/Pages/Calculator.xcplaygroundpage/Sources/AbstractOperation.swift
@@ -9,6 +9,6 @@ public protocol AbstractOperation {
     
     var symbol: String { get }
     
-    func operate(_ firstInput: Int, with secondInput: Int) -> Int
+    func operate(_ firstInput: Double, with secondInput: Double) -> Double
     
 }
